Refactor Code
 * Add Google Services files to gitignore.
 * Create sub-directories to group related files.
 * Remove prefix My- from classes names.
 * Remove redundant comments.
 * Remove test directory.
 * Rename all files using snake case.
 * Split up sign up screens.
 * Use lowercase for package name.
