+++ .metadata
@@ -1,7 +1,5 @@
-# This file tracks properties of this Flutter project.
-# Used by Flutter tool to assess capabilities and perform upgrades etc.
-#
-# This file should be version controlled and should not be manually edited.
+# This file should be version controlled
+# and should not be manually edited.
 
 version:
   revision: 2d2a1ffec95cc70a3218872a2cd3f8de4933c42f
