+++ lib/models/user.dart
@@ -8,8 +8,10 @@ class User {
   String sharedKey;
 
   static final User _user = User._internal();
+
   factory User() {
     return _user;
   }
+
   User._internal();
 }
